+++ tests/test_gpipe.py
@@ -468,16 +468,16 @@ def test_named_children():
         model.a
 
 
-def test_recommend_balancing():
-    with pytest.raises(ValueError, match='balancing'):
+def test_recommend_auto_balance():
+    with pytest.raises(ValueError, match='torchgpipe.balance'):
         # balance is required
         GPipe(nn.Sequential())
 
-    with pytest.raises(ValueError, match='balancing'):
+    with pytest.raises(ValueError, match='torchgpipe.balance'):
         # module and sum of balance have differen length (module: 0, sum of balance: 1)
         GPipe(nn.Sequential(), [1])
 
-    with pytest.raises(ValueError, match='balancing'):
+    with pytest.raises(ValueError, match='torchgpipe.balance'):
         # module and sum of balance have different length (module: 2, sum of balance: 1)
         GPipe(nn.Sequential(nn.Linear(1, 1), nn.Linear(1, 1)), [1])