+++ torchgpipe_balancing.py
@@ -1,2 +1,2 @@
-# torchgpipe_balancing has moved to torchgpipe.balancing in v0.0.5.
-raise ImportError('import torchgpipe.balancing instead')
+# 'torchgpipe_balancing' has moved to 'torchgpipe.balance' in v0.0.5.
+raise ImportError("import 'torchgpipe.balance' instead")